+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `tt cluster`: worker publish subcommand.
 - `tt cluster`: worker show subcommand.
+- `tt cluster`: worker delete subcommand.
 
 ### Changed